renaming robot initialization
