+++ tests/compilation_tests/robot_test/robot_test.py
@@ -2,7 +2,7 @@ import copy
 import math
 import unittest
 
-from engine.robot_logic.robot_initialization import Robot
+from engine.robot_logic.robot import Robot
 from engine.robot_logic.traversal import travel
 from engine.robot_state import Robot_State
