+++ tests/functionality_tests/obstacle_tracking_test.py
@@ -1,6 +1,6 @@
 import unittest
 import math
-from engine.robot_logic.robot_initialization import Robot
+from engine.robot_logic.robot import Robot
 from engine.robot_logic.set_up import track_obstacle
 from engine.robot_state import Robot_State
 from engine.mission import Mission
