+++ tests/functionality_tests/setup_phase_test.py
@@ -1,4 +1,4 @@
-from engine.robot_logic.robot_initialization import Robot
+from engine.robot_logic.robot import Robot
 from engine.robot_logic.set_up import execute_setup
 from engine.phase import Phase
 from engine.robot_state import Robot_State
