+++ tests/functionality_tests/test_on_line.py
@@ -1,5 +1,5 @@
 import unittest
-from engine.robot_logic.robot_initialization import Robot
+from engine.robot_logic.robot import Robot
 from engine.robot_state import Robot_State
 import numpy as np
